Update PriceProductRepository.php
Revert order of conditions in applyCriteria method. It is needed because when we pass $priceProductCriteriaTransfer, which is call in Spryker\Zed\PriceProduct\Business\Model\Product\PriceProductConcreteWriter::executePersistProductConcretePrice and there transfer object is decorate with productAbstractId. This cause that the first condition is always true and only delete prices for abstract. This change fix this and delete for both concrete and abstract